Complete the task using lists
Above issue resolved.
