--- src/main.py
@@ -1,22 +1,0 @@
-'''
-Product:
-Description:
-Author:
-'''
-from generator import listClass
-
-class Solution():
-    def __init__(self):
-        inst = listClass()
-        self.listOfNumbers = inst.number_generator()
-    
-    def your_solution(self):
-        '''
-        Write you solution in here!
-        '''
-        
-        pass
-
-if __name__ == "__main__":
-    inst = Solution()
-    inst.your_solution()
